% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled.rbd.R
\name{pooled.rcbd}
\alias{pooled.rcbd}
\title{Pooled RCBD}
\usage{
pooled.rcbd(loc, treat, block, Y)
}
\arguments{
\item{loc}{A vector of location (factor)}

\item{treat}{A vector of treatments in each experiment and over the locations (factor)}

\item{block}{A vector of blocks in each experiment and over the locations (factor)}

\item{Y}{A vector of output variable (numeric)}
}
\value{
An ANOVA table of combined analysis
}
\description{
Provides an ANOVA table for combined analysis of series of experiments conducted in RCBD over the locations of time
}
\examples{
pooled.rcbd(loc=)
}
\references{

}
